add all tests

+++ src/main/java/api/AllProductsGet.java
@@ -6,13 +6,13 @@ public class AllProductsGet {
     private String article;
     private String id;
     private String category;
-    private Integer price;
+    private Double price;
     private Integer qty;
     private String insertedAt;
     private String last_qty_changed;
     private String currency;
 
-    public AllProductsGet(String name, String article, String id, String category, Integer price, Integer qty, String insertedAt, String last_qty_changed, String currency) {
+    public AllProductsGet(String name, String article, String id, String category, Double price, Integer qty, String insertedAt, String last_qty_changed, String currency) {
         this.name = name;
         this.article = article;
         this.id = id;
@@ -56,7 +56,7 @@ public class AllProductsGet {
         return category;
     }
 
-    public Integer getPrice() {
+    public Double getPrice() {
         return price;
     }
 
@@ -92,7 +92,7 @@ public class AllProductsGet {
         this.category = category;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
